src/requirements.txt
appdirs==1.4.4
asgiref==3.5.0
attrs==21.4.0
boto3==1.21.46
botocore==1.24.46
coverage==6.3.2
Django==4.0.4
django-crispy-forms==1.14.0
django-storages==1.12.3
factory-boy==3.2.1
Faker==13.7.0
gunicorn==20.1.0
inflection==0.5.1
iniconfig==1.1.1
jmespath==1.0.0
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.8
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-factoryboy==2.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
